codex: inline executor http state into handler
